+++ game_lib/src/cards/mod.rs
@@ -1,1 +1,6 @@
-pub mod model;
+pub mod properties;
+pub(crate) mod serialization;
+pub mod types;
+pub mod world;
+
+pub mod errors;