+++ game_lib/src/cards/properties/fix_modifier.rs
@@ -1,0 +1,19 @@
+use serde::{Deserialize, Serialize};
+use crate::cards::world::resources::Resources;
+
+#[derive(Serialize, Deserialize, Clone, Debug)]
+#[serde(rename_all = "camelCase")]
+pub enum FixModifier {
+    Increase(Resources),
+    Decrease(Resources),
+}
+
+impl FixModifier {
+    pub fn value(&self) -> isize {
+        match self {
+            FixModifier::Increase(r) => r.value().clone() as isize,
+            FixModifier::Decrease(r) => -1 * r.value().clone() as isize
+        }
+    }
+}
+