+++ seccardgame/src/migrations/mod.rs
@@ -1,2 +1,2 @@
 pub mod version_one;
-pub mod version_two;
+pub mod version_three;